dao/src/test/java/com/epam/brest/dao/CategoryDaoJDBCImplTest.java
package com.epam.brest.dao;

import com.epam.brest.model.Category;
import java.util.Collections;
import java.util.List;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.ArgumentCaptor;
import org.mockito.ArgumentMatchers;
import org.mockito.Captor;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Mockito;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.jdbc.core.RowMapper;
import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
import org.springframework.jdbc.core.namedparam.SqlParameterSource;

import static org.mockito.ArgumentMatchers.any;

@ExtendWith(MockitoExtension.class)
public class CategoryDaoJDBCImplTest {

    @InjectMocks
    private CategoryDaoJDBCImpl categoryDaoJDBC;

    @Mock
    private NamedParameterJdbcTemplate namedParameterJdbcTemplate;

    @Captor
    private ArgumentCaptor<RowMapper<Category>> captorMapper;

    @Captor
    private ArgumentCaptor<SqlParameterSource> captorSource;

    @AfterEach
    public void check() {
        Mockito.verifyNoMoreInteractions(namedParameterJdbcTemplate);
    }

    @Test
    public void testFindAllCategories() {

        Category category = new Category();
        List<Category> list = Collections.singletonList(category);

        Mockito.when(namedParameterJdbcTemplate.query(any(), ArgumentMatchers.<RowMapper<Category>>any()))
                .thenReturn(list);

        List<Category> result = categoryDaoJDBC.findAllCategories();

        Mockito.verify(namedParameterJdbcTemplate).query(any(), captorMapper.capture());

        RowMapper<Category> mapper = captorMapper.getValue();

        Assertions.assertNotNull(mapper);

        Assertions.assertNotNull(result);
        Assertions.assertFalse(result.isEmpty());
        Assertions.assertSame(category, result.get(0));
    }

    @Test
    public void testGetCategoryById() {
        int id = 0;
        Category category = new Category();

        Mockito.when(namedParameterJdbcTemplate.queryForObject(
                any(),
                ArgumentMatchers.<SqlParameterSource>any(),
                ArgumentMatchers.<RowMapper<Category>>any())
        ).thenReturn(category);

        Category result = categoryDaoJDBC.getCategoryById(id);

        Mockito.verify(namedParameterJdbcTemplate)
                .queryForObject(any(), captorSource.capture(), captorMapper.capture());

        SqlParameterSource source = captorSource.getValue();
        RowMapper<Category> mapper = captorMapper.getValue();

        Assertions.assertNotNull(source);
        Assertions.assertNotNull(mapper);

        Assertions.assertNotNull(result);
        Assertions.assertSame(category, result);
    }
}
